Organize code
Create new package and create new Objects
