+++ Project02Eclipse/src/com/hemebiotech/analytics/Main.java
@@ -1,0 +1,8 @@
+package com.hemebiotech.analytics;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+    }
+}
