+++ Project02Eclipse/src/com/hemebiotech/analytics/analytics/AnalyticsCounter.java
@@ -1,0 +1,37 @@
+package com.hemebiotech.analytics.analytics;
+
+import java.io.BufferedReader;
+import java.io.FileReader;
+import java.io.FileWriter;
+
+public class AnalyticsCounter {
+	private static int headacheCount = 0;	// initialize to 0
+	private static int rashCount = 0;		// initialize to 0
+	private static int pupilCount = 0;		// initialize to 0
+	
+	public static void main(String[] args) throws Exception {
+		// first get input
+		BufferedReader reader = new BufferedReader (new FileReader("symptoms.txt"));
+		String line = reader.readLine();
+
+		while (line != null) {
+			if (line.equals("headache")) {
+				headacheCount++;
+			}
+			else if (line.equals("rush")) {
+				rashCount++;
+			}
+			else if (line.contains("pupils")) {
+				pupilCount++;
+			}
+			line = reader.readLine();	// get another symptom
+		}
+		
+		// next generate output
+		FileWriter writer = new FileWriter ("result.out");
+		writer.write("headache: " + headacheCount + "\n");
+		writer.write("rash: " + rashCount + "\n");
+		writer.write("dialated pupils: " + pupilCount + "\n");
+		writer.close();
+	}
+}
