+++ Project02Eclipse/src/com/hemebiotech/analytics/analytics/counter/SymptomCounter.java
@@ -1,0 +1,4 @@
+package com.hemebiotech.analytics.analytics.counter;
+
+public class SymptomCounter {
+}
