+++ Project02Eclipse/src/com/hemebiotech/analytics/analytics/reader/SymptomReader.java
@@ -1,0 +1,49 @@
+package com.hemebiotech.analytics.analytics.reader;
+
+import com.hemebiotech.analytics.analytics.reader.ISymptomReader;
+
+import java.io.BufferedReader;
+import java.io.FileReader;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Simple brute force implementation
+ *
+ */
+public class SymptomReader implements ISymptomReader {
+
+	private final String filepath;
+	
+	/**
+	 * 
+	 * @param filepath a full or partial path to file with symptom strings in it, one per line
+	 */
+	public SymptomReader(String filepath) {
+		this.filepath = filepath;
+	}
+	
+	@Override
+	public List<String> getSymptoms() {
+		ArrayList<String> result = new ArrayList<String>();
+		
+		if (filepath != null) {
+			try {
+				BufferedReader reader = new BufferedReader (new FileReader(filepath));
+				String line = reader.readLine();
+				
+				while (line != null) {
+					result.add(line);
+					line = reader.readLine();
+				}
+				reader.close();
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+		}
+		
+		return result;
+	}
+
+}
