+++ Project02Eclipse/src/com/hemebiotech/analytics/analytics/writer/ISymptomWriter.java
@@ -1,0 +1,4 @@
+package com.hemebiotech.analytics.analytics.writer;
+
+public interface ISymptomWriter {
+}
